Relocate extension lifecycle regressions
